Change InvokeStateHelper.clone(InvokeState) to not clone call stack, as cloning the call stack is incompatible with how the webMethods Adapter Runtime (ART) releases connections, and results in leaked connections.
ART relies on an com.wm.app.b2b.server.invoke.IServiceCompletionListener
object to release an adapter connection, and unfortunately the method
com.wm.app.b2b.server.invoke.ServiceCompletionImpl.serviceCompleted(boolean)
which notifies listeners of the completion of a service, only supports
top-level service completion. So even if a child service is the service
which calls the ART adapter service that opens the connection, the
connection is not released until the top-level invoking service
completes.

The permafrost.tundra.server.ServerThreadPoolExecutor object provides
a convenient way to use multiple threads to process a work load
concurrently. However it was using InvokeStateHelper.clone(InvokeState)
to initialize each ServerThread in the thread pool with a clone of
a given InvokeState (typically the top-level invoking service's
InvokeState). Previous to this change the call stack was also being
cloned as part of the InvokeState clone (the cloned call stack
included all calls up to the top-level service). Therefore, when a
ServerThread in the ServerThreadPoolExecutor thread pool is assigned
a service to invoke, it is invoked with a call stack that already
includes at least one other service (the service which constructed
the thread pool and assigned work to it), and so when the service
invoked by the thread completes the serviceCompleted(boolean)
method checks the size of the call stack, and since it indicates
that the thread invoked service is not the top-level service it
does not notify the ART service completion listener object to release
the connection and instead the connection leaks.

This change fixes the ART connection leaks by not cloning the call
stack when cloning an InvokeState used to initialize a ServerThread.