Change SubstitutionHelper to support a new type of substitution for invoking a service dynamically and substituting in the returned output.
For example, a variable containing the following substitution string:

    %invoke:example.employee:get/name?id=1%

Will invoke the service `example.employee:get` with the input `id`
set to the value `1`, and substitute the value of the returned output
`name` into the variable.

A service invoke substitution is performed if an invoke URI is
specified. An invoke URI is an opaque URI with the following
components:

    scheme = invoke
    body   = service name/output pipeline variable
    query  = input pipeline variables